feat: added the FROM_THE_ENTRY_RELATION recipient type
This recipient type add the ability to send notification to
email address that is on a relation of he entry.
For example :
Suppose we have a collection named `Planet` which has a column `discorveredBy` which is a relation to the `User` table. This type of recipient will enable us, for example, to send an e-mail to the person who discovered the planet by retrieving his e-mail from the `discorveredBy` relationship.